updated tests with struct member datatypes
used updated code generator to generate tests that have struct member datatypes specified when possible

+++ test/petstore_v2/petstore/src/api_StoreApi.jl
@@ -35,7 +35,7 @@ function deleteOrder(_api::StoreApi, response_stream::Channel, in_orderId::Int64
 end
 
 function _oacinternal_getInventory(_api::StoreApi; _mediaType=nothing)
-    _ctx = OpenAPIClient.Ctx(_api.client, "GET", Dict{String, Int32}, "/store/inventory", ["api_key", ])
+    _ctx = OpenAPIClient.Ctx(_api.client, "GET", Dict{String, Int64}, "/store/inventory", ["api_key", ])
     OpenAPIClient.set_header_accept(_ctx, ["application/json", ])
     OpenAPIClient.set_header_content_type(_ctx, (_mediaType === nothing) ? [] : [_mediaType])
     return _ctx
@@ -47,7 +47,7 @@ Returns a map of status codes to quantities
 
 Params:
 
-Return: Dict{String, Int32}
+Return: Dict{String, Int64}
 """
 function getInventory(_api::StoreApi; _mediaType=nothing)
     _ctx = _oacinternal_getInventory(_api; _mediaType=_mediaType)
